Fixed issues where ASP.NET controls were too large for the Bootstrap areas allocated to them; added comments and touched up for final version

+++ MPAS/Admin/DeleteUser.aspx.cs
@@ -32,35 +32,37 @@ namespace MPAS.Admin
         }
         protected void DeleteButton_Click(object sender, EventArgs e)
         {
-            using (var _db = new MPAS.Models.ApplicationDbContext())
+            if (IsValid)
             {
-                string studentNumber = StudentNumberDelete_TextBox.Text;
-                var myItem = (from c in _db.Users where c.UserName == studentNumber select c).FirstOrDefault();
-                if (myItem != null)
+                using (var _db = new MPAS.Models.ApplicationDbContext())
                 {
-                    _db.Users.Remove(myItem);
-                    _db.SaveChanges();
+                    string studentNumber = StudentNumberDelete_TextBox.Text;
+                    var myItem = (from c in _db.Users where c.UserName == studentNumber select c).FirstOrDefault();
+                    if (myItem != null)
+                    {
+                        _db.Users.Remove(myItem);
+                        _db.SaveChanges();
 
-                    _db.Database.ExecuteSqlCommand("DELETE FROM ProfileDetails WHERE StudentNumber=@p0", studentNumber);
+                        _db.Database.ExecuteSqlCommand("DELETE FROM ProfileDetails WHERE StudentNumber=@p0", studentNumber);
 
-                    // indicate success
-                    StatusLabel.Text = "User deleted";
-                    StatusLabel.Visible = true;
-                    StatusLabel.ForeColor = System.Drawing.Color.Green;
+                        // indicate success
+                        StatusLabel.Text = "User deleted";
+                        StatusLabel.Visible = true;
+                        StatusLabel.ForeColor = System.Drawing.Color.Green;
 
-                    StudentNumberDelete_TextBox.Text = "";
-                    //Reload the table when user is deleted
-                    LoadUserTable();
-                }
-                else
-                {
-                    //indicate failure
-                    StatusLabel.Text = "Unable to locate user";
-                    StatusLabel.Visible = true;
-                    StatusLabel.ForeColor = System.Drawing.Color.Red;
+                        StudentNumberDelete_TextBox.Text = "";
+                        //Reload the table when user is deleted
+                        LoadUserTable();
+                    }
+                    else
+                    {
+                        //indicate failure
+                        StatusLabel.Text = "Unable to locate user";
+                        StatusLabel.Visible = true;
+                        StatusLabel.ForeColor = System.Drawing.Color.Red;
+                    }
                 }
             }
-
         }
 
         protected void StudentNumberValidate(object source, ServerValidateEventArgs args)
@@ -68,6 +70,7 @@ namespace MPAS.Admin
             args.IsValid = (Regex.IsMatch(args.Value, @"^[a-zA-Z]{6,6}[0-9]{3,3}$") // student number
                 || Regex.IsMatch(args.Value, @"^[0-9]{8,8}$") // staff number
                 );
+            args.IsValid = args.IsValid && args.Value != "01360406";
         }
     }
